footer integration
