+++ src/App.jsx
@@ -8,6 +8,7 @@ import LoginForm from "./pages/LoginRegister/LoginForm";
 import Register from "./pages/LoginRegister/RegisterForm";
 import Home from "./pages/Home/HomePage";
 import Navbar from "./components/Navbar";
+import Footer from "./components/Footer";
 import ProductDetails from "./pages/ProductDetails/Details";
 import ProductSearch from "./pages/ProductsPage/SearchPage";
 import Details from "./pages/ProductDetails/DetailsCard";
@@ -19,7 +20,6 @@ import AddProduct from "./pages/MerchantDashboard/AddEditProduct";
 import MerchantDashboard from "./pages/MerchantDashboard/Dashboard";
 import AddressPage from "./pages/Address/AddressPage";
 import MerchantForm from "./pages/MerchantForm";
-import Footer from "./pages/footer";
 import "./App.css";
 import "./index.css";
 
@@ -32,13 +32,25 @@ const AppLayout = () => (
 	</>
 );
 
+const WithFooter = () => (
+	<>
+		<main>
+			<Outlet />
+		</main>
+		<Footer />
+	</>
+);
+
 const App = () => {
 	return (
 		<Router>
 			<Routes>
 				<Route element={<AppLayout />}>
-					<Route path="/" element={<Home />} />
-					<Route path="/merchant" element={<MerchantDashboard />} />
+					<Route element={<WithFooter />}>
+						<Route path="/" element={<Home />} />
+						<Route path="/merchant" element={<MerchantDashboard />} />
+						<Route path="/search" element={<ProductSearch />} />
+					</Route>
 					<Route path="/merchantregistration" element={<MerchantForm />} />
 					<Route path="/home" element={<Home />} />
 					<Route path="/details" element={<Details />} />
@@ -46,7 +58,7 @@ const App = () => {
 					<Route path="/products/add" element={<AddProduct />} />
 					<Route path="/products/update/:id" element={<AddProduct />} />
 					<Route path="/receipt/:id" element={<PurchaseReceipt />} />
-					<Route path="/search" element={<ProductSearch />} />
+
 					<Route path="/products" element={<ProductSearch />} />
 					<Route path="/cart" element={<Cart />} />
 					<Route path="/purchases" element={<Purchases />} />
